backend/pyproject.toml
[tool.poetry]
name = "createathon-backend"
version = "0.1.0"
description = "Backend for the Createathon Interactive Creator Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8.1"
django = "^4.2"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
markdown = "^3.5"
django-cors-headers = "^4.3"
channels = "^4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
black = "^23.12"
isort = "^5.13"
flake8 = "^6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
